# internal/modules/store/storages/gen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package db

import (
	"database/sql"

	"github.com/google/uuid"
)

type Category struct {
	ID          uuid.UUID
	Name        string
	Slug        string
	Description sql.NullString
	CreatedAt   int64
	UpdatedAt   sql.NullInt64
}

type Product struct {
	ID          uuid.UUID
	Name        string
	Slug        string
	Description sql.NullString
	Price       float64
	CategoryID  uuid.UUID
	CreatedAt   int64
	UpdatedAt   sql.NullInt64
}